# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AndMI API/Open API System",
        "description": "Andijon Mashinasozlik Instituti Ochiq API tizimi",
        "contact": {
            "email": "andmiatmdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/files": {
            "get": {
                "tags": [
                    "Файлы"
                ],
                "summary": "Получение списка всех файлов",
                "operationId": "f8d4504d5ab4240c97fda7ace991333e",
                "responses": {
                    "200": {
                        "description": "Список файлов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Мой документ"
                                            },
                                            "slug": {
                                                "type": "string",
                                                "example": "moi-dokument"
                                            },
                                            "author_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "types_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "format_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "group_of": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "part_of": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "used_times": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "zip": {
                                                "type": "string",
                                                "example": "my-document.zip"
                                            },
                                            "active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Файлы"
                ],
                "summary": "Создание файла",
                "operationId": "7228cc192e1dab9ab20ec46a83dffd5e",
                "requestBody": {
                    "description": "Данные для создания файла",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "slug",
                                    "author_id",
                                    "user_id",
                                    "types_id",
                                    "format_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Мой документ"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "moi-dokument"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "types_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "format_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "group_of": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "part_of": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "used_times": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "zip": {
                                        "type": "string",
                                        "example": "my-document.zip"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Файл успешно создан"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "422": {
                        "description": "Некорректные данные"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/files/{id}": {
            "get": {
                "tags": [
                    "Файлы"
                ],
                "summary": "Получение файла",
                "operationId": "f5493ba67a9e53f1ac1ed9f57b9162f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID файла",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Файл найден"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Файл не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Файлы"
                ],
                "summary": "Обновление файла",
                "operationId": "f8b282a84287c8a904c20a8f3a57301c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID файла",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления файла",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "slug",
                                    "author_id",
                                    "user_id",
                                    "types_id",
                                    "format_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Мой обновленный документ"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "moi-obnovlennyi-dokument"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "types_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "format_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "group_of": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "part_of": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "used_times": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "zip": {
                                        "type": "string",
                                        "example": "my-updated-document.zip"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Файл успешно обновлен"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Файл не найден"
                    },
                    "422": {
                        "description": "Некорректные данные"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Файлы"
                ],
                "summary": "Удаление файла",
                "operationId": "aec67941f2c518b4d87e831e86781065",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID файла",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Файл успешно удален"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Файл не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token orqali autentifikatsiyadan o'tish",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}